Add `import` contacts feature.
Currently, users have to move their file to the /data folder and rename it to addressbook.json, which is inefficient and prone to errors when telemarkers have many different files for their different shifts.

Add an `Import` button under `File` in the top menu of the GUI which opens file chooser to select file to import.

Using the GUI to select files in this way is more intuitive than CLI as it is similar to other applications that open files.