Create HistoryManager class
Teletubbies does not keep track of entered inputs.

This makes it difficult for the user to trace through inputs entered previously. Users also have to type out entire command again if small typo was made.

As a step towards implementing the history command, let's create a generic HistoryManager class that can store the list of states of a type T.