Create CommandInputHistory class using HistoryManager and perform integration tests
Teletubbies does not keep track of entered inputs.

This makes it difficult for the user to trace through inputs entered previously. Users also have to type out entire command again if small typo was made.

As a step towards implementing the history command, let's create an abstraction for this, CommandInputHistory, using HistoryManager. Let's also integrate the CommandInputHistory with ModelManager.