Create HistoryCommand and integrate adding inputs with LogicManager
Teletubbies does not keep track of entered inputs.

This makes it difficult for the user to trace through inputs entered previously. Users also have to type out entire command again if small typo was made.

As a step towards implementing the history command, let's create a HistoryCommand and integrate the tracking of user inputs with LogicManager and Model.

Let's also teach Parser to parse HistoryCommand